Answer changes due to c5 OS update on 2024/10/16
We think these answers change due to a change in compiler update intel/2022 to intel/2023
that coincided with a OS update on gaea c5. We have not been able to verify that this
is the only reason for the answer changes but since gnu and nvidia (formerly pgi) did
not change, we are 51% confident it is due to the compiler version.